Insert new transactions into a secondary DDB table
Up till now, only a single table was used with partitions "year-month" + "balance".
This would make it difficult to query in order to get a year's worth of balance data
(12 calls for each year-month or a scan, which is bad).

Instead, we create a second table that has "year" + "month-day" with a single balance
value (the lowest balance that day). This way we can query on "year" and plot
month-day + balance values, which are already in sorted order.