Change TransactionInserter to TransactionHandler
The TransactionInserter only contained logic to insert transactions into the
transaction datastore. The TransactionHandler also contains logic to delete
transactions from the datastore.

The balance is now stripped of its decimals for transactions - this prevents
small changes in bank balance (due to inconsistencies in the bank csv's) from
duplicating entries in the datastore.